requirements.txt
aws-cdk-lib==2.29.1
constructs>=10.0.0,<11.0.0
black==22.10.0
boto3==1.26.25
click==8.1.3
pre-commit==2.20.0
PyYAML==6.0
pytest==7.2.0
pytest-dotenv==0.5.2
moto==4.1.7
coverage==7.2.3
ffmpeg==1.4
yt-dlp==2024.4.9
pytube==15.0.0
chat_downloader==0.2.8
pandas==2.2.2
numpy==1.26.4